refactor: extract management helpers from AgentWorkspace
